Revert moving user profiles
