+++ plugin/src/main/java/net/dzikoysk/funnyguilds/user/NPCUserProfile.java
@@ -1,4 +1,4 @@
-package net.dzikoysk.funnyguilds.user.profile;
+package net.dzikoysk.funnyguilds.user;
 
 public class NPCUserProfile extends FakeUserProfile implements UserProfile {
     public NPCUserProfile() {
