+++ plugin/src/main/java/net/dzikoysk/funnyguilds/user/UserManager.java
@@ -8,9 +8,6 @@ import java.util.UUID;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.stream.Collectors;
 import net.dzikoysk.funnyguilds.FunnyGuilds;
-import net.dzikoysk.funnyguilds.user.profile.FakeUserProfile;
-import net.dzikoysk.funnyguilds.user.profile.NPCUserProfile;
-import net.dzikoysk.funnyguilds.user.profile.UserProfile;
 import org.apache.commons.lang3.Validate;
 import org.bukkit.OfflinePlayer;
 import org.bukkit.entity.Player;
